backend/requirements_processfile.txt
accelerate==0.28.0
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
blinker==1.4
boilerpy3==1.0.7
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6

cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.3
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.4.0
google-ai-haystack==0.2.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
haystack-ai==2.0.0
haystack-bm25==1.0.2
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.22.1
idna==3.6
importlib-metadata==4.6.4
install==1.3.5
jeepney==0.7.1
Jinja2==3.1.3
jsonformer==0.12.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-imports==0.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
openai==1.14.3
packaging==23.2
pandas==2.2.1
pillow==10.2.0
posthog==3.5.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pyngrok==7.1.6
pyparsing==2.4.7
pypdfium2==4.28.0
pytesseract==0.3.10
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==5.4.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
ssh-import-id==5.11
starlette==0.36.3
streamlit==1.32.2
sympy==1.12
systemd-python==234
tenacity==8.2.3
termcolor==2.4.0
toml==0.10.2
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
uvicorn==0.29.0
wadllib==1.3.6
watchdog==4.0.0
zipp==1.0.0
